requirements.txt
Django==1.5.1
South>=0.7.6
django-annoying>=0.7.7
django-bootstrap-toolkit==2.8.0
django-debug-toolbar==0.9.4
django-extensions==1.0.3
psycopg2>=2.5
gunicorn-console
msgpack-python==0.3.0
pytz==2012j
django-tastypie==0.9.12
jsonfield==0.9.7

# celery
celery==3.0.16
django-celery==3.0.11